# internal/svc/iam-jwks/api/generated.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
)

// Defines values for InvalidAttributeSource.
const (
	Body   InvalidAttributeSource = "body"
	Header InvalidAttributeSource = "header"
	Path   InvalidAttributeSource = "path"
	Query  InvalidAttributeSource = "query"
)

// Valid indicates whether the value is a known member of the InvalidAttributeSource enum.
func (e InvalidAttributeSource) Valid() bool {
	switch e {
	case Body:
		return true
	case Header:
		return true
	case Path:
		return true
	case Query:
		return true
	default:
		return false
	}
}

// Defines values for JWKAlg.
const (
	EdDSA JWKAlg = "EdDSA"
)

// Valid indicates whether the value is a known member of the JWKAlg enum.
func (e JWKAlg) Valid() bool {
	switch e {
	case EdDSA:
		return true
	default:
		return false
	}
}

// Defines values for JWKCrv.
const (
	Ed25519 JWKCrv = "Ed25519"
)

// Valid indicates whether the value is a known member of the JWKCrv enum.
func (e JWKCrv) Valid() bool {
	switch e {
	case Ed25519:
		return true
	default:
		return false
	}
}

// Defines values for JWKKty.
const (
	OKP JWKKty = "OKP"
)

// Valid indicates whether the value is a known member of the JWKKty enum.
func (e JWKKty) Valid() bool {
	switch e {
	case OKP:
		return true
	default:
		return false
	}
}

// Defines values for JWKUse.
const (
	Sig JWKUse = "sig"
)

// Valid indicates whether the value is a known member of the JWKUse enum.
func (e JWKUse) Valid() bool {
	switch e {
	case Sig:
		return true
	default:
		return false
	}
}

// Error defines model for Error.
type Error struct {
	Code              string              `json:"code"`
	InvalidAttributes *[]InvalidAttribute `json:"invalid_attributes,omitempty"`
	Message           string              `json:"message"`
	Name              string              `json:"name"`
	OperationId       string              `json:"operation_id"`
}

// InvalidAttribute defines model for InvalidAttribute.
type InvalidAttribute struct {
	Path   string                 `json:"path"`
	Reason string                 `json:"reason"`
	Rule   string                 `json:"rule"`
	Source InvalidAttributeSource `json:"source"`
}

// InvalidAttributeSource defines model for InvalidAttribute.Source.
type InvalidAttributeSource string

// JWK RFC 8037. EdDSA/Ed25519 only.
type JWK struct {
	Alg JWKAlg `json:"alg"`
	Crv JWKCrv `json:"crv"`
	Kid string `json:"kid"`
	Kty JWKKty `json:"kty"`
	Use JWKUse `json:"use"`

	// X base64url-encoded Ed25519 public key.
	X string `json:"x"`
}

// JWKAlg defines model for JWK.Alg.
type JWKAlg string

// JWKCrv defines model for JWK.Crv.
type JWKCrv string

// JWKKty defines model for JWK.Kty.
type JWKKty string

// JWKUse defines model for JWK.Use.
type JWKUse string

// JWKS defines model for JWKS.
type JWKS struct {
	Keys []JWK `json:"keys"`
}

// InternalServerError defines model for InternalServerError.
type InternalServerError = Error

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetJWKS the current key set
	// (GET /.well-known/jwks.json)
	GetJWKS(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// GetJWKS the current key set
// (GET /.well-known/jwks.json)
func (_ Unimplemented) GetJWKS(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetJWKS operation middleware
func (siw *ServerInterfaceWrapper) GetJWKS(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetJWKS(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/.well-known/jwks.json", wrapper.GetJWKS)
	})

	return r
}

type InternalServerErrorJSONResponse Error

type GetJWKSRequestObject struct {
}

type GetJWKSResponseObject interface {
	VisitGetJWKSResponse(w http.ResponseWriter) error
}

type GetJWKS200JSONResponse JWKS

func (response GetJWKS200JSONResponse) VisitGetJWKSResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetJWKS500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response GetJWKS500JSONResponse) VisitGetJWKSResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetJWKS the current key set
	// (GET /.well-known/jwks.json)
	GetJWKS(ctx context.Context, request GetJWKSRequestObject) (GetJWKSResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetJWKS operation middleware
func (sh *strictHandler) GetJWKS(w http.ResponseWriter, r *http.Request) {
	var request GetJWKSRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetJWKS(ctx, request.(GetJWKSRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetJWKS")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetJWKSResponseObject); ok {
		if err := validResponse.VisitGetJWKSResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}